Reorganize and rename packages
