+++ src/agents/dcop/adopt/receiveMessageBehaviour.java
@@ -1,4 +1,4 @@
-package agent;
+package agents.dcop.adopt;
 
 import jade.core.Agent;
 import jade.core.behaviours.CyclicBehaviour;
