+++ src/agents/dcop/preinit/searchForChildrenBehaviour.java
@@ -1,4 +1,4 @@
-package agent;
+package agents.dcop.preinit;
 
 import jade.core.Agent;
 import jade.core.behaviours.WakerBehaviour;
@@ -8,22 +8,22 @@ import jade.domain.FIPAAgentManagement.DFAgentDescription;
 import jade.domain.FIPAAgentManagement.ServiceDescription;
 import models.DCOPAgentData;
 
-public class searchForLowerNeighboursBehaviour extends WakerBehaviour {
+public class searchForChildrenBehaviour extends WakerBehaviour {
 	
-	private static final long serialVersionUID = 2788042325314110781L;
+	private static final long serialVersionUID = -3081227894323949053L;
 	DCOPAgentData data = new DCOPAgentData();
 	
-    public searchForLowerNeighboursBehaviour(Agent a, long period, DCOPAgentData data) {
+    public searchForChildrenBehaviour(Agent a, long period, DCOPAgentData data) {
         super(a, period);
         this.data = data;
     }
 
     @Override
     protected void onWake() {
-        for (String lowerNeighbourName : data.getLowerNeighboursNames()) {
+        for (String childName : data.getChildrenNames()) {
             DFAgentDescription template = new DFAgentDescription();
             ServiceDescription serviceTemplate = new ServiceDescription();
-            serviceTemplate.setType("agent-" + lowerNeighbourName);
+            serviceTemplate.setType("agent-" + childName);
             template.addServices(serviceTemplate);
 
             DFAgentDescription []resultSearch = null;
@@ -35,10 +35,10 @@ public class searchForLowerNeighboursBehaviour extends WakerBehaviour {
             }
 
             if (resultSearch.length != 0) {
-                data.setLowerNeighbour(resultSearch[0].getName());
-                System.out.println("[NEIGH      ] Registering agent " +
-                        resultSearch[0].getName().getLocalName() +
-                        " as " + myAgent.getLocalName() + "'s lower neighbour");
+                data.setChild(resultSearch[0].getName());
+                System.out.println("[CHILD      ] Registering agent " +
+                                    resultSearch[0].getName().getLocalName() +
+                                    " as " + myAgent.getLocalName() + "'s child");
             }
         }
     }
