+++ src/agents/instantiator/InstantiatorAgent.java
@@ -1,4 +1,4 @@
-package instantiator;
+package agents.instantiator;
 
 import graph.Graph;
 import graph.Node;
@@ -44,7 +44,7 @@ public class InstantiatorAgent extends Agent {
 		//Launch nodes as agents on platform
         for(Node node : graph.getNodes()) {
         	try {
-                getContainerController().createNewAgent(node.getID(), "agent.DCOPAgent", node.getArgs()).start();
+                getContainerController().createNewAgent(node.getID(), "agents.dcop.DCOPAgent", node.getArgs()).start();
             } catch (StaleProxyException stale) {
             	System.out.println("Could not start agent\n");
                 stale.printStackTrace();
