+++ src/messages/AdoptMessage.java
@@ -6,4 +6,5 @@ public abstract class AdoptMessage implements Serializable, MessageTypes{
 	public static final long serialVersionUID = 247808931907196589L;
 			
 	public abstract Integer getMessageType();
+	public abstract String toString();
 }
